requirements.txt
asgiref==3.11.1
click==8.3.1
colorama==0.4.6
cssbeautifier==1.15.4
Django==6.0.2
djlint==1.36.4
EditorConfig==0.17.1
jsbeautifier==1.15.4
json5==0.13.0
pathspec==1.0.4
PyYAML==6.0.3
regex==2026.1.15
six==1.17.0
sqlparse==0.5.5
tqdm==4.67.3
tzdata==2025.3
python-dotenv==1.1.1
